requirements.txt
boto3==1.35.86
Pillow>=10.0.0
python-magic>=0.4.27
moviepy>=1.0.3
botocore>=1.31.0
gradio==4.44.1
